requirements.txt
pika==1.2.0
PyYAML==5.4.1
packaging
robonomics-interface
importlib-resources